+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 
 Feature:
 
-* add support folder
+* add `support` folder
   * move some class to support folder
   * add Action class for executable action
   * add some general Action classes
+* add `utils.validateParameter` method
 
 Improvement: